requirements.txt
amqp==2.6.1
asgiref==3.2.10
autopep8==1.5.4
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.7.0rc1
Django==3.1
django-3-jet==1.0.8
django-celery-beat==2.0.0
django-ckeditor==5.9.0
django-cors-headers==3.5.0
django-js-asset==1.2.2
django-timezone-field==4.0
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
oauthlib==3.1.0
Pillow==7.2.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
stripe==2.51.0
toml==0.10.1
Unidecode==1.1.1
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
